dockercompose: move helper functions around
